requirements.txt
streamlit>=1.35.0
google-genai>=0.3.0
pillow>=10.3.0